backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
google-generativeai==0.3.2
python-dotenv==1.0.0
resend==0.8.0
slowapi==0.1.9
email-validator==2.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
